refactor clean code
