+++ pkg/proxy/dnsserver/dnsserver_test.go
@@ -1,4 +1,4 @@
-package daemon
+package dnsserver
 
 import (
 	"testing"
